% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_variable_gene.R
\name{cluster_spatial_expr_pattern}
\alias{cluster_spatial_expr_pattern}
\title{Cluster spatial expression features}
\usage{
cluster_spatial_expr_pattern(
  obj,
  assay = "enhanced",
  feature,
  bin = 20,
  n.cluster = 12,
  cluster.method = "ward.D2",
  ncol.plot = 4,
  n.core = 1
)
}
\arguments{
\item{obj}{seurat object}

\item{assay}{assay to use}

\item{feature}{features (genes) to get spatial expression patterns}

\item{bin}{number of bins used to segregate the image}

\item{n.cluster}{number of clusters}

\item{cluster.method}{hclust method}

\item{ncol.plot}{number of columns for plotting}

\item{n.core}{number of CPU cores used}
}
\value{
return a vector showing the genes in each cluster
}
\description{
Clustering the genes based on the spatial expression pattern.
}
