% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_validation}
\alias{cross_validation}
\title{10 fold cross validation}
\usage{
cross_validation(spRNA, scRNA)
}
\arguments{
\item{spRNA}{seurat object of spatial transcriptome}

\item{scRNA}{seurat object of scRNA-seq}
}
\value{
correlation between inferred spatial expression and ground true.
}
\description{
10 fold cross validation
}
