% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_variable_gene.R
\name{get_spatial_features}
\alias{get_spatial_features}
\title{Get spatial expression features}
\usage{
get_spatial_features(obj, assay, feature, bin, n.core = 1)
}
\arguments{
\item{obj}{seurat object}

\item{assay}{assay to use}

\item{feature}{features (genes) to get spatial expression patterns}

\item{bin}{number of bins used to segregate the image}

\item{n.core}{number of CPU cores used}
}
\value{
return a matrix. row: gene, column: spatial expression features
}
\description{
Get spatial expression features
}
