% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_variable_gene.R
\name{get_spatial_features_image}
\alias{get_spatial_features_image}
\title{get spatial expression features in each image}
\usage{
get_spatial_features_image(obj, assay, feature, img, bin, n.core = 1)
}
\arguments{
\item{obj}{seurat object}

\item{assay}{assay to use}

\item{feature}{features (genes) to get spatial expression patterns}

\item{img}{choose image}

\item{bin}{number of bins used to segregate the image}

\item{n.core}{number of CPU cores used}
}
\value{
return a matrix. row: gene, column: spatial expression features
}
\description{
get spatial expression features in each image
}
