% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSpatial.R
\name{infer_Hierarchy}
\alias{infer_Hierarchy}
\title{Hierarchy version of iSpatial}
\usage{
infer_Hierarchy(
  spRNA,
  scRNA,
  dims = 1:30,
  k.neighbor = 30,
  cluster.resolution = 0.1,
  infered.assay = "enhanced",
  weighted.KNN = TRUE,
  RNA.weight = 0.3,
  n.core = 10,
  correct.spRNA = TRUE,
  correct.scRNA = FALSE,
  correct.neighbor = 3
)
}
\description{
Hierarchy version of iSpatial
}
