% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSpatial.R
\name{iSpatial}
\alias{iSpatial}
\title{Main function of iSpatial}
\usage{
iSpatial(
  spRNA,
  scRNA,
  dims = 1:30,
  k.neighbor = 30,
  infered.assay = "enhanced",
  RNA.weight = NA,
  n.core = 10,
  correct.expr = TRUE,
  correct.neighbor = 5
)
}
\arguments{
\item{spRNA}{seurat object of spatial transcriptome data}

\item{scRNA}{seurat object of single cell RNA-seq data}

\item{dims}{which dimensions to use when find the nearest neighbors}

\item{k.neighbor}{number of neighbors to use when infer the expression}

\item{infered.assay}{names of output assay in seurat object}

\item{RNA.weight}{the weight of scRNA-seq expression when genes detected both
in spRNA and scRNA. When choose defaulted "NA", we use a dynamics weight
to assign scRNA values to spRNA. The dynamics weight based on the correlation
between scRNA cell and spRNA cell. If a RNA.weight (from 0 to 1) were given,
the inferred expression = (1 - RNA.weight) * spRNA + RNA.weight * scRNA.}

\item{n.core}{number of CPU cores used to parallel.}

\item{correct.expr}{Whether to stabilize expression in scRNA and spRNA.}

\item{correct.neighbor}{number of nearest neighbors used to correct expr.}
}
\value{
returns a seurat object with inferred expression in infered.assay.
}
\description{
This function integrates scRNA-seq and spatial transcriptome. With limited
genes in merFISH/SeqFISH, iSpatial infer/enhance all genes from scRNA-seq.
Finally, this function generates spatially expression of all genes.
}
