% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSpatial.R
\name{iSpatial_Hierarchy}
\alias{iSpatial_Hierarchy}
\title{Hierarchy version of iSpatial}
\usage{
iSpatial_Hierarchy(
  spRNA,
  scRNA,
  dims = 1:30,
  n.core = 10,
  RNA.weight = 0.3,
  cluster.resolution = 0.1,
  infered.assay = "enhanced"
)
}
\description{
Hierarchy version of iSpatial
}
