% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSN_merFISH}
\alias{MSN_merFISH}
\title{merFISH of mouse nucleus accumbens}
\format{
An object of class Seurat
}
\source{
\url{https://www.nature.com/articles/s41593-021-00938-x}
}
\usage{
MSN_merFISH
}
\description{
An object of class Seurat
253 features across 12984 samples within 1 assay
Active assay: RNA (253 features, 0 variable features)
Cell locations store in MSN_merFISH@images
}
\keyword{datasets}
