% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSN_scRNA}
\alias{MSN_scRNA}
\title{scRNA-seq (10x platform) of mouse nucleus accumbens}
\format{
An object of class Seurat
}
\source{
\url{https://www.nature.com/articles/s41593-021-00938-x}
}
\usage{
MSN_scRNA
}
\description{
An object of class Seurat
26145 features across 10000 samples within 1 assay
Active assay: RNA (26145 features, 1052 variable features)
}
\keyword{datasets}
