% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NA_scRNA}
\alias{NA_scRNA}
\title{scRNA-seq (10x platform) of mouse nucleus accumbens}
\format{
An object of class Seurat
}
\source{
\url{https://www.nature.com/articles/s41593-021-00938-x}
}
\usage{
NA_scRNA
}
\description{
An object of class Seurat.
26145 features across 6000 samples within 1 assay
}
\keyword{datasets}
