% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSpatial.R
\name{recommend_k}
\alias{recommend_k}
\title{function of iSpatial}
\usage{
recommend_k(
  spRNA,
  scRNA,
  dims = 1:30,
  k.neighbor = c(5, seq(10, 80, by = 10)),
  infered.assay = "enhanced",
  n.core = 8,
  correct.spRNA = FALSE,
  correct.scRNA = FALSE,
  correct.weight.NN = 0.2,
  correct.neighbor = 5
)
}
\arguments{
\item{spRNA}{seurat object of spatial transcriptome data}

\item{scRNA}{seurat object of single cell RNA-seq data}

\item{dims}{which dimensions to use when find the nearest neighbors}

\item{k.neighbor}{a list of K to test}

\item{infered.assay}{names of output assay in seurat object}

\item{n.core}{number of CPU cores used to parallel.}

\item{correct.spRNA}{Whether to stabilize expression in spRNA.}

\item{correct.scRNA}{Whether to stabilize expression in scRNA.}

\item{correct.neighbor}{number of nearest neighbors used to correct expr.}

\item{weighted.KNN}{we use a dynamics weight to assign scRNA values to spRNA.
The dynamics weight based on the correlation between scRNA cell and spRNA
cell.}

\item{RNA.weight}{the weight of scRNA-seq expression when genes detected both
in spRNA and scRNA. RNA.weight should be 0 to 1.
The inferred expression = (1 - RNA.weight) * spRNA + RNA.weight * scRNA.}
}
\value{
returns a ggplot2 object.
}
\description{
recommand optimized K value for KNN
}
