% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSpatial.R
\name{run_harmony}
\alias{run_harmony}
\title{run harmony integration}
\usage{
run_harmony(seurat_obj, genes_select)
}
\arguments{
\item{seurat_obj}{merged seurat object}

\item{genes_select}{genes used to integration, usually targeted genes.}
}
\value{
seurat object
}
\description{
This function calls harmony package(https://github.com/immunogenomics/harmony)
to integrate the scRNA-seq and spatial transcriptome data.
}
