% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cor.R
\name{sample_region}
\alias{sample_region}
\title{Random sample spatial regions}
\usage{
sample_region(obj, bin = 20)
}
\arguments{
\item{obj}{seurat object}

\item{bin}{number of bins used to segregate the image}
}
\value{
A vector including the cell names in a random sampling region
}
\description{
Get the cell names from random sampling spatial regions
}
