% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cor.R
\name{spatial_cor}
\alias{spatial_cor}
\title{Calculate spatially expression correlation}
\usage{
spatial_cor(obj1, assay1, obj2, assay2, feature = NA, bin = 20, n.sample = 300)
}
\arguments{
\item{obj1}{seurat object 1}

\item{assay1}{choose assay for obj1}

\item{obj2}{seurat object 2}

\item{assay2}{choose assay for obj2}

\item{feature}{select features to calulate}

\item{bin}{number of bins used to segregate the image}

\item{n.sample}{how many times for sampling regions}
}
\value{
Spatial expression correlation of each gene
}
\description{
Calculate spatially expression correlation between two objects. This function
can be used to validation the performance of inferred data.
}
