% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_variable_gene.R
\name{spatial_var_gene_slice}
\alias{spatial_var_gene_slice}
\title{Find spatial variable genes}
\usage{
spatial_var_gene_slice(obj, assay, img, bin, n.core = 1)
}
\arguments{
\item{obj}{seurat object}

\item{assay}{select assay}

\item{img}{select image}

\item{bin}{number of bins used to segregate the image}

\item{n.core}{number of CPU cores used}
}
\value{
return a matrix wtih spatial variable genes
}
\description{
Find spatial variable genes
}
