% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSpatial.R
\name{stabilize_expr}
\alias{stabilize_expr}
\title{stabilize expression}
\usage{
stabilize_expr(obj, neighbor = 5, npcs = 10, n.core = 10)
}
\arguments{
\item{obj}{seurat object}

\item{neighbor}{number of nearest neighbors used to correct}

\item{npcs}{number of principal component (PC) used to RunPCA}

\item{n.core}{number of CPU cores used to parallel}
}
\value{
seurat object with corrected expression
}
\description{
This function use nearest neighbor to correct the expression.
This could remove some  False Positive / False negative expressions.
This function corrects "RNA" assay "data" slot.
}
