% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_validation}
\alias{cross_validation}
\title{10 fold cross validation}
\usage{
cross_validation(
  spRNA,
  scRNA,
  fold = 10,
  correct.spRNA = TRUE,
  correct.scRNA = FALSE
)
}
\arguments{
\item{spRNA}{seurat object of spatial transcriptome}

\item{scRNA}{seurat object of scRNA-seq}

\item{fold}{number of X fold crossing validation}

\item{correct.spRNA}{Whether to stabilize expression in spRNA.}

\item{correct.scRNA}{Whether to stabilize expression in scRNA.}
}
\value{
correlation between inferred spatial expression and ground true.
}
\description{
10 fold cross validation
}
