% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_variable_gene.R
\name{spatial_signature_plot}
\alias{spatial_signature_plot}
\title{Plot spatial expression}
\usage{
spatial_signature_plot(
  object,
  assay = "enhanced",
  gene.set,
  set.name,
  pt.size.factor = 2,
  alpha = 0.5,
  color.scale = c("gray95", "gray80", "pink", "orange", "red", "red4"),
  max.cutoff = NA,
  legend = FALSE
)
}
\arguments{
\item{object}{seurat object}

\item{assay}{choose assay}

\item{gene.set}{vector contain the genes to plot}

\item{set.name}{set the name for the plot}

\item{pt.size.factor}{point size}

\item{alpha}{transparency of the points}

\item{color.scale}{set the color scale. Default:
c("gray95", "gray80", "pink", "orange", "red", "red4")}

\item{max.cutoff}{set max of the expression value.
Default:  0.99\% of the max.}

\item{legend}{whether to plot the legend}
}
\value{
ggplot object
}
\description{
Plot the spatial expression patterns on a group of genes.
}
