% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_variable_gene.R
\name{spatial_variable_genes}
\alias{spatial_variable_genes}
\title{Find spatial variable genes}
\usage{
spatial_variable_genes(obj, assay, bin = 20, n.core = 1)
}
\arguments{
\item{obj}{seurat object}

\item{assay}{select assay}

\item{bin}{number of bins used to segregate the image}

\item{n.core}{number of CPU cores used}
}
\value{
return a matrix wtih spatial variable genes
}
\description{
Bin the images into small chunks. For each gene, detected the numbers of
expression cells in each chunk, comparing the distribution with random.
}
